start debugging.... :) / :(

+++ include/Coin.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "StaticObject.h"
-
+#include "Utillities.h"
 
 class Coin : public StaticObject
 {
@@ -12,7 +12,7 @@ public:
 	int coinValue();
 	virtual void print(sf::RenderWindow& window);
 	//virtual sf::Sprite getSprite();
-	virtual void setLocation(const int y, const int x);
+	virtual void setLocation(const float y, const float x);
 
 private:
 	sf::Sprite m_CoinPng;
