+++ include/Enemy.h
@@ -1,13 +1,14 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "MovingObject.h"
+#include "Utillities.h"
 
 class Enemy : public MovingObject
 {
 public:
     Enemy(sf::Texture picture, const int speed);
     virtual ~Enemy() = default;
-    virtual void setLocation(const int y, const int x);
+    virtual void setLocation(const float y, const float x);
     void moveLocation(const sf::Vector2f& d, sf::Time t);
     virtual void print(sf::RenderWindow& window);
     virtual void move(const sf::Time& time);//, const sf::Vector2f &direction); // empy
