+++ include/Floor.h
@@ -1,6 +1,10 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "StaticObject.h"
+#include "Utillities.h"
+#include<SFML/Graphics/Sprite.hpp>
+#include <iostream>
+#include <ostream>
 
 class Floor : public StaticObject
 {
@@ -9,7 +13,7 @@ public:
 	virtual ~Floor() = default;
 	virtual void print(sf::RenderWindow & window);
 	//virtual sf::Sprite getSprite();
-	virtual void setLocation(const int y, const int x);
+	virtual void setLocation(const float y, const float x);
 
 private:
 	sf::Sprite m_floorPng;
